refactor: replace Double with BigDecimal for monetary fields across investment DTOs and use cases

+++ src/main/java/com/jf/PetApp/application/investment/dto/AllocationSliceDTO.java
@@ -2,9 +2,14 @@ package com.jf.PetApp.application.investment.dto;
 
 import com.jf.PetApp.core.domain.enums.InvestmentType;
 
+import java.math.BigDecimal;
+
+/**
+ * Money fields are {@link BigDecimal} — see docs/BACKEND_MODULE_PLAN.md §12.
+ */
 public record AllocationSliceDTO(
     InvestmentType type,
-    Double currentValue,
-    Double portfolioPercent
+    BigDecimal currentValue,
+    BigDecimal portfolioPercent
 ) {
 }
